@@ -15,14 +15,14 @@
   ],
   "devDependencies": {
     "parcel-bundler": "^1.12.4",
-    "polotno": "^0.43.12",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "polotno": "^1.0.0-2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "author": "Anton Lavrenov",
   "license": "MIT",
   "dependencies": {
     "chrome-aws-lambda": "^10.1.0",
-    "puppeteer": "^15.1.1"
+    "puppeteer": "^15.4.0"
   }
 }
